+++ sdk/storage/Azure.Storage.Files.DataLake/CHANGELOG.md
@@ -1,7 +1,7 @@
 # Release History
 
 ## 12.3.0-preview.2 (Unreleased)
-
+- Fixed bug where DataLakeUriBuilder would return LastDirectoryOrFileName and DirectoryOrFilePath URL-encoded.
 
 ## 12.3.0-preview.1 (2020-07-03)
 - Added support for service version 2019-12-12.